% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.predict.gamlss.R
\name{ds.predict.gamlss}
\alias{ds.predict.gamlss}
\title{Predictions for Generalized Additive Models for Location Scale and Shape}
\usage{
ds.predict.gamlss(object, newdata, what = "mu", type = "link")
}
\arguments{
\item{object}{a DataSHIELD GAMLSS fitted model output by \code{ds.gamlss}.}

\item{newdata}{a data frame containing new values for the explanatory variables used
in the model.}

\item{what}{distribution parameter that should be predicted. The default is 
\code{what="mu"}.}

\item{type}{the default value \code{type="link"} gets the linear predictor for the 
specified distribution parameter. \code{type="response"} gets the fitted values for
the distribution parameter.}
}
\value{
a vector with the predictions for \code{newdata}.
}
\description{
This is the \code{ds.gamlss} method which produces predictions for a new 
data set for a specfied parameter from a \code{ds.gamlss} object.
}
\details{
The predict function assumes that the object given in \code{newdata} is a 
data frame containing the right x-variables used in the model.
}
\author{
Annika Swenne
}
