% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsGamlssClient-package.R
\docType{package}
\name{dsGamlssClient-package}
\alias{dsGamlssClient}
\alias{dsGamlssClient-package}
\title{dsGamlssClient: DataSHIELD client-side functions for Generalized Additive Models for Location, Scale and Shape}
\description{
DataSHIELD client-side functions to fit Generalized Additive Models for Location, Scale and Shape using DataSHIELD. DataSHIELD is a software infrastructure which allows you to do non-disclosive federated analysis on sensitive data. Note that DataSHIELD has a client-server infrastructure, so the dsGamlss package needs to be used in conjunction with the dsGamlssClient package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/dsGamlssClient}
  \item \url{https://bips-hb.github.io/dsGamlssClient/}
  \item Report bugs at \url{https://github.com/bips-hb/dsGamlssClient/issues}
}

}
\author{
\strong{Maintainer}: Annika Swenne \email{swenne@leibniz-bips.de}

}
\keyword{internal}
