% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPooledMean.R
\name{getPooledMean}
\alias{getPooledMean}
\title{Gets a pooled statistical mean from the servers}
\usage{
getPooledMean(datasources, x)
}
\arguments{
\item{datasources}{A list of \code{\link[DSI]{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link[DSI]{datashield.connections_default}}.}

\item{x}{A string with the name of a numeric vector for which the mean should be computed}
}
\value{
A numeric value, giving the pooled mean.
}
\description{
This is an internal function based on the internal \code{\link[dsBaseClient]{getPooledMean}} function from \code{dsBaseClient} (version 6.3.0).
}
\details{
This function is called to avoid calling the client function \code{\link[dsBaseClient]{ds.mean}}
which may stop the process due to some checks not required when computing a mean inside
a function.
}
\keyword{internal}
