% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDefined.R
\name{isDefined}
\alias{isDefined}
\title{Check if an object is defined on all servers}
\usage{
isDefined(datasources = NULL, obj = NULL, error.message = TRUE)
}
\arguments{
\item{datasources}{A list of \code{\link[DSI]{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link[DSI]{datashield.connections_default}}.}

\item{obj}{A string with the name of the object(s) to look for.}

\item{error.message}{A Boolean which specifies if the function should stop and return
an error message when the input object is not defined on one or more servers (\code{error.message=TRUE}) 
or if it should return a list of \code{TRUE/FALSE} indicating on which server the object is defined
(\code{error.message=FALSE}). Default \code{error.message=TRUE}.}
}
\value{
An error message if \code{error.message} argument is set to \code{TRUE} (default)
and if the input object is not defined on one or more servers, or a Boolean value if
\code{error.message=FALSE}.
}
\description{
This is an internal function based on the internal \code{\link[dsBaseClient]{isDefined}} function from \code{dsBaseClient} (version 6.3.0).
}
\details{
In DataSHIELD an object included in an analysis must be defined (i.e. exists) on all servers. If not the process should halt.
}
\author{
Demetris Avraam for DataSHIELD Development Team
}
\keyword{internal}
