% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssChecks.R
\name{gamlssChecks}
\alias{gamlssChecks}
\title{Checks if the elements in the gamlss model have the right characteristics}
\usage{
gamlssChecks(
  formula,
  sigma.formula,
  nu.formula,
  tau.formula,
  data,
  datasources
)
}
\arguments{
\item{formula}{a character, a regression formula given as a string character}

\item{sigma.formula}{a character, a regression formula given as a string character}

\item{nu.formula}{a character, a regression formula given as a string character}

\item{tau.formula}{a character, a regression formula given as a string character}

\item{data}{a character, the name of an optional data frame containing the variables in
in the \code{formula}.}

\item{datasources}{a list of \code{\link[DSI]{DSConnection-class}} objects obtained after login. If the <datasources>
the default set of connections will be used: see \link[DSI]{datashield.connections_default}.}
}
\value{
an integer 0 if check was passed and 1 if failed
}
\description{
This is an internal function required by the client function \code{\link{ds.gamlss}}
to verify all the variables and ensure the process does not halt inadvertanly.
}
\details{
the variables are checked to ensure they are defined, not empty (i.e. are not missing
at complete).
}
\author{
Gaye, A.
}
\keyword{internal}
